kickstarter-workflow/requirements.txt
# Kickstarter Monitor Workflow Dependencies

# HTTP 请求
requests>=2.31.0

# 数据验证
pydantic>=2.0.0

# 日期处理
python-dateutil>=2.8.0

# 环境变量
python-dotenv>=1.0.0

# 日志
structlog>=23.0.0

# 类型检查
typing-extensions>=4.0.0
